+++ apps/web/playwright.config.ts
@@ -29,7 +29,7 @@ export default defineConfig({
     },
   ],
   webServer: {
-    command: 'bun start',
+    command: 'bun dev',
     url: 'http://localhost:3000',
     reuseExistingServer: !process.env['CI'],
   },